pyproject.toml
[project]
name = "protein-design"
version = "0.1.0"
description = "ESM2 evotuning pipeline for antibody design"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "transformers",
    "datasets",
    "biopython",
    "pandas",
    "wandb",
    "pyyaml",
    "tqdm",
    "python-dotenv",
    "numpy",
    "requests",
    "matplotlib",
    "scipy",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/protein_design"]
